feature: Create helm charts for kubernetes deployment. Touches #4504 (#4)
Creates two helm charts for on-premises. One is for service only and the second
for fast deployment of the whole stack required to run cloud services. Every
chart has its own README.md file with information about installation and how to
verify the deployment with tests.

Stack chart uses the service's chart as a dependency as well as bitnami/mysql
and bitnami/redis charts. It's insecure and not persistent by default, there is
a warning with that information added to the readme file. It can be used for
local development in minikube as well, the only requirement is having access to
CKEditor's container registry and license key of on-premises.

Touches: cksource/cs#4504.