feature: Create Azure Terraform modules. Closes: cksource/cs#8996 (#8)
Create two terraform modules to provision the CKEditor Collaboration
Server on Microsoft Azure cloud:
- infrastructure
- service

The infrastructure module provisions everything the service needs to
run. It creates a virtual network with subnets for each element of
infrastructure. Provisions relational database as a MySQL Flexible
Server and Redis cluster, as well as Azure Storage Account for blob
storage. Kubernetes cluster with Application Gateway is used as an
environment to run the CKEditor Collaboration Server container and
expose it to the public network.

The service module handles provisioning the CKEditor Collaboration
Server deployment by using the Helm chart. It can be omitted while
provisioning and the Helm chart can be used directly instead.

Example usage of those modules is stored in a separate catalog. It can
be seamlessly used for fast provisioning whole infrastructure and the
CKEditor Collaboration Server.

Closes: cksource/cs#8996