Create helm charts for CKBox On-Premises
Create deployment configuration for Kubernetes environments for CKBox
On-Premises. There are two options to deploy the CKBox. Production ready
helm chart for installing only the backend of CKBox On-Premises and a
development stack to run the CKBox On-premises with all necessary
components - MySQL and Redis.

Provisioning script for development stack is based on the same script
from CKEditor Collaboration Server development stack, as they are
similar in the terms of infrastructure requirements.

Closes: cksource/cs#17707